fs: fix reference counting
The scenario was that Create starts an inode with ref=2. Then the file
is closed so release causes ref=1. Then an unlink happens and the inode
is freed. Then the inode kernel cache is freed that causes forget to run
and the inode isn't there.

The solution is that open/release don't need to account for the internal
kernel cache reference count, so the release for closing after create
don't decrease the ref count.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>