ibv: fix ibv launcher but change local_mode setup
This fixes the ibv launcher, but moves gasnet_init to the beginning of main().
This has the affect that one needs to use the GASNet launcher (e.g. udp) even
when using local mode. We'll fix this later either by 1) building two
executables or 2) allowing gasnet_init to fail and not throwing an error in
the case that it turns out the program was launched with local mode.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>